backend/requirements.txt
# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0

redis==5.0.1
rq==1.16.2

# HTTP Client
httpx>=0.26.0,<0.28.0

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# Development Tools
ruff==0.8.4
black==24.10.0
pre-commit==4.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Environment
python-dotenv==1.0.1

# CORS
python-multipart==0.0.18

# AI/ML for RAG implementation
openai==1.54.4
pypdf>=3.9.0
transformers==4.53.0
torch==2.8.0
sentence-transformers==3.0.1
langgraph>=0.2.0

# Database
asyncpg==0.29.0
supabase==2.9.1

# File handling
aiofiles==24.1.0